executeSwap
